requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.1.2
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.3.0
backoff==2.2.1
backports.asyncio.runner==1.2.0
banks==2.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==44.0.0
customtkinter==5.2.1
darkdetect==0.8.0
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.118.0
fastembed==0.7.3
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.5.0
fsspec==2025.9.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
griffe==1.14.0
grpcio==1.75.1
h11==0.14.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.8.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.2
jiter==0.8.2
joblib==1.4.2
jq==1.8.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.78
langchain-ollama==0.3.10
langchain-openai==0.2.14
langchain-postgres==0.0.12
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.4
langdetect==1.0.9
langsmith==0.4.32
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.4
llama-index-cli==0.5.3
llama-index-core==0.14.4
llama-index-embeddings-fastembed==0.5.0
llama-index-embeddings-ollama==0.8.3
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-openai==0.6.4
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-qdrant==0.8.6
llama-index-workflows==2.7.1
llama-parse==0.6.54
loguru==0.7.3
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.23.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
ollama==0.6.0
onnxruntime==1.23.1
openai==1.109.1
openpyxl==3.1.5
orjson==3.10.13
packaging==24.2
pandas==2.2.3
parso==0.8.4
pgvector==0.2.5
pgxnclient==1.3.2
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
portalocker==3.2.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.26.1
protobuf==4.25.1
psutil==6.1.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyarrow==14.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.0
pydantic-settings==2.7.0
pydantic_core==2.41.1
pydeck==0.8.1b0
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==5.1.0
pyreadline3==3.5.4
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Python-IO==0.3
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.15.1
RapidFuzz==3.11.0
referencing==0.31.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.13.1
rsa==4.9.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.48.0
streamlit==1.28.2
striprtf==0.0.26
sympy==1.14.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
toolz==0.12.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.6.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.2
unstructured==0.16.11
unstructured-client==0.28.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.37.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
win32_setctime==1.2.0
wrapt==1.17.0
yarl==1.18.3
zipp==3.17.0
zstandard==0.25.0
